Add shared types into package dir
