+++ src/packages/shared-types/index.ts
@@ -1,3 +1,1 @@
-export type ExampleType = {
-  testing: string;
-};
+export * from "./issue";
