+++ src/services/ui/package.json
@@ -49,6 +49,7 @@
     "tailwindcss": "^3.3.1",
     "typescript": "^4.9.3",
     "vite": "^4.2.0",
-    "vitest": "^0.30.1"
+    "vitest": "^0.30.1",
+    "shared-types": "*"
   }
 }
