+++ src/services/ui/src/api/index.ts
@@ -1,4 +1,3 @@
 export * from "./useCreateIssue";
 export * from "./useGetIssue";
 export * from "./useGetIssues";
-export * from "./validators";
